build: port build system to cmake
This isn't totally done yet but I've successfully built the os and all
programs with the new system, but there are some problems.

First is the need to compilation of targets using two different
toolchains. Right now, the idea is to have to separate cmake builds and
run the native one first. This hopefully can be solved by a recursive
cmake invocation or something.
Second is the need for installation of libc and libm before anything
will link successfully. This is fixed by manually running the
install-libc target before a build. It follows that programs also don't
respect changes to libc properly, when they should have to relink.
Thirdly, none of the helper Makefile things are ported, so you can't
make a disk image or the os iso file yet.

Presumably these problems will be fixed with some build script that
orechestrates all of these unfortunate circumanstances, which should be
well worth it considering how much faster builds are (with ninja).

+++ libs/libc/include/sys/select.h
@@ -30,11 +30,11 @@ int select(int numfds, fd_set *__restrict readfds, fd_set *__restrict writefds, 
 
 #define FD_ISSET(fd, set) ((set)->fds[fd / sizeof(unsigned char) / __CHAR_BIT__] & (1U << (fd % (sizeof(unsigned char) * __CHAR_BIT__))))
 
-#define FD_ZERO(set)                                                                  \
-    do {                                                                              \
-        for (int i = 0; i < FD_SETSIZE / sizeof(unsigned char) / __CHAR_BIT__; i++) { \
-            (set)->fds[i] = 0;                                                        \
-        }                                                                             \
+#define FD_ZERO(set)                                                                            \
+    do {                                                                                        \
+        for (unsigned long i = 0; i < FD_SETSIZE / sizeof(unsigned char) / __CHAR_BIT__; i++) { \
+            (set)->fds[i] = 0;                                                                  \
+        }                                                                                       \
     } while (0)
 
 #endif /* _SYS_SELECT_H */
